algos/4 sem/5 lab/code/clustering_project/requirements.txt
numpy
scikit-learn
matplotlib
cure-cluster
pandas
PyYAML
